[WIP] Rename models package
