+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/db/entities/Product.java
@@ -1,4 +1,4 @@
-package com.OBJ2100.ExamApp.db.models;
+package com.OBJ2100.ExamApp.db.entities;
 
 public class Product {
 	private int quantityInStock;
